chore(docs): consolidate data files and trim example frontmatter
- Move docs-sidebar.ts and docs-meta.json under src/data/ alongside examples.ts
- Strip redundant title/description/slug/tags/last_reviewed from example-post
  frontmatter (already owned by src/data/examples.ts); narrow the examplePosts
  schema to a single optional `image` field
- Drop the trailing period appended to example page hero titles
