+++ docs/src/components/Sidebar.astro
@@ -1,5 +1,5 @@
 ---
-import { sidebar, type SidebarItem } from '../docs-sidebar';
+import { sidebar, type SidebarItem } from '../data/docs-sidebar';
 
 export interface Props {
   /** Slug of the currently-rendered doc, e.g. `getting_started/quickstart`. */
