+++ docs/src/content.config.ts
@@ -30,19 +30,14 @@ const docs = defineCollection({
 
 // Example walkthroughs — ported from github.com/cocoindex-io/examples.
 // One .md file per slug in src/content/example-posts, rendered by
-// src/pages/examples/[slug].astro beneath the shared hero.
+// src/pages/examples/[slug].astro beneath the shared hero. Card metadata
+// (title, description, tags, etc.) lives in src/data/examples.ts — the
+// filename drives the slug.
 const examplePosts = defineCollection({
   loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/example-posts' }),
   schema: z
     .object({
-      title: z.string(),
-      description: z.string().optional(),
-      slug: z.string(),
       image: z.string().optional(),
-      tags: z.array(z.string()).optional(),
-      // YAML parses ISO dates into Date; accept either so hand-edited
-      // string dates also work.
-      last_reviewed: z.union([z.string(), z.date()]).optional(),
     })
     .passthrough(),
 });
