+++ docs/src/content/example-posts/image_search.md
@@ -1,10 +1,5 @@
 ---
-title: Image Search App with ColPali and FastAPI
-description: 'Build image search index with ColPali and FastAPI'
-slug: image_search
 image: https://cocoindex.io/blobs/docs/img/examples/image_search/cover.png
-tags: [vector-index, multi-modal]
-last_reviewed: 2026-01-18
 ---
 
 [→ View on GitHub](https://github.com/cocoindex-io/cocoindex/tree/main/examples/image_search)
